Updated db, but not code
Added notes of what to do

In an element, all its attributes can be listed, but the bounds, whitelisting and conditions apply to if it can be read or written.

Added some quick lookup flags in the hords to make it faster to figure out what to do when there is a read or write there, or when map or shape needs to be considered, or when time needs to be considered per attribute.

 All user permissions are now only in the types, the types also have the sum map and shape for that time period, and gives the current period of time information for quicker lookup (which is a sum of its attributes)

 Attributes now have the bounds and not the rules, the time and location attributes do not affect each other. The location bounds never change due to time (it's too complicated otherwise)

 Rules now only do the affinities when being added to a set, when in a set, and when designing types. Rules do not change based on map or time in that rule (map and time bounds now no longer in the rules). An attribute being off does disable the rule checks.

 If a type has some checks turned off because of in a set, those rules for joining a child set will still be off because its in the parent set context