Removed z order; retirement only on types
shapes can block events, but this is not designed for organizing parallel events on unrelated stuff. So cannot apply

when publishing, that is when all the child types and attribute parents are checked

Attributes are not retired anymore, just their owning types.
If the owning type is retired, all attributes there cannot be used as parents to new attributes

Changed some names for clarity and make sure unique to later search in code