Added concept of copy/pasting live types
This way, element can get values changes in the live attr, then pass that on as a group to targets.

The copied live type is still on the original, and the copy values are independent after that. The source can remove that live type without affecting the others it copied too.

This allows lego'sh and other behavior that can be quite cool.

The api and scope of this project is now defined in the act and evt

All go through the things and use paths as the source and target, with or without rules