Changes to concept live requirements, phase, energy,charge
energy and charge too complex

energy is just rate limiting, which can be done in second layer

charge can have many of the same logic using much simpler requirements based on the type. if you think of types being minus or positive in what is there -  about the same thing

And all the types can match up, easier to see and debug. Charges added a new layer that was going to be in the way and more importantly, not used so much

Added phase because things need to help evolve